refactor: change dir names to match main component in dir
