+++ src/web/src/views/workspace/WSEditorCommandContent/CommandDeleteDialog.tsx
