+++ src/web/src/views/workspace/WSEditorCommandContent/OutputCard.tsx
