+++ src/web/src/views/workspace/WSEditorCommandContent/OutputDialog.tsx
