+++ src/web/src/views/workspace/WSEditorCommandContent/WSEditorCommandContent.tsx
