+++ src/web/src/views/workspace/WSEditorCommandContent/index.ts
@@ -1,2 +1,2 @@
-export { default as WSEditorCommandContent, DecodeResponseCommand } from "./WSEditorCommandContent";
+export { default, DecodeResponseCommand } from "./WSEditorCommandContent";
 export type { Command, Resource, ResponseCommand, ResponseCommands, Example, Plane } from "../interfaces";
