+++ src/web/src/views/workspace/WSEditorCommandGroupContent/CommandGroupDeleteDialog.tsx
