+++ src/web/src/views/workspace/WSEditorCommandGroupContent/WSEditorCommandGroupContent.tsx
