+++ src/web/src/views/workspace/WSEditorCommandGroupContent/index.tsx
